+++ WebApiTerra1000/Controllers/NomenclatureController.cs
@@ -48,6 +48,22 @@ namespace WebApiTerra1000.Controllers
 
         [AllowAnonymous]
         [HttpGet()]
+        [Route("api/nomenclature_preview/")]
+        public ContentResult GetNomenclaturePreview(string code, int id, FormatType format = FormatType.Xml)
+        {
+            return Controller.RunTask(new Task<ContentResult>(() =>
+            {
+                string response = string.IsNullOrEmpty(code)
+                    ? TerraUtils.Sql.GetResponse<string>(SessionFactory, SqlQueries.GetNomenclatureFromIdPreview, new SqlParameter("id", id))
+                    : TerraUtils.Sql.GetResponse<string>(SessionFactory, SqlQueries.GetNomenclatureFromCodePreview, new SqlParameter("code", code));
+                XDocument xml = XDocument.Parse(response ?? $"<{TerraConsts.Goods} />", LoadOptions.None);
+                XDocument doc = new(new XElement(TerraConsts.Response, xml.Root));
+                return BaseSerializeEntity<XDocument>.GetResult(format, doc, HttpStatusCode.OK);
+            }), format);
+        }
+
+        [AllowAnonymous]
+        [HttpGet()]
         [Route("api/nomenclatures/")]
         public ContentResult GetNomenclatures(DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 10,
             FormatType format = FormatType.Xml)
@@ -64,12 +80,43 @@ namespace WebApiTerra1000.Controllers
 
         [AllowAnonymous]
         [HttpGet()]
+        [Route("api/nomenclatures_preview/")]
+        public ContentResult GetNomenclaturesPreview(DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 10,
+            FormatType format = FormatType.Xml)
+        {
+            return Controller.RunTask(new Task<ContentResult>(() =>
+            {
+                string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, SqlQueries.GetNomenclaturesPreview,
+                    TerraUtils.Sql.GetParameters(startDate, endDate, offset, rowCount));
+                XDocument xml = XDocument.Parse(response ?? $"<{TerraConsts.Goods} />", LoadOptions.None);
+                XDocument doc = new(new XElement(TerraConsts.Response, xml.Root));
+                return BaseSerializeEntity<XDocument>.GetResult(format, doc, HttpStatusCode.OK);
+            }), format);
+        }
+
+        [AllowAnonymous]
+        [HttpGet()]
         [Route("api/nomenclaturescosts/")]
         public ContentResult GetNomenclaturesCosts(DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 10, FormatType format = FormatType.Xml)
         {
             return Controller.RunTask(new Task<ContentResult>(() =>
             {
                 string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, SqlQueries.GetNomenclaturesCosts,
+                    TerraUtils.Sql.GetParameters(startDate, endDate, offset, rowCount));
+                XDocument xml = XDocument.Parse(response ?? $"<{TerraConsts.Goods} />", LoadOptions.None);
+                XDocument doc = new(new XElement(TerraConsts.Response, xml.Root));
+                return BaseSerializeEntity<XDocument>.GetResult(format, doc, HttpStatusCode.OK);
+            }), format);
+        }
+
+        [AllowAnonymous]
+        [HttpGet()]
+        [Route("api/nomenclaturescosts_preview/")]
+        public ContentResult GetNomenclaturesCostsPreview(DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 10, FormatType format = FormatType.Xml)
+        {
+            return Controller.RunTask(new Task<ContentResult>(() =>
+            {
+                string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, SqlQueries.GetNomenclaturesCostsPreview,
                     TerraUtils.Sql.GetParameters(startDate, endDate, offset, rowCount));
                 XDocument xml = XDocument.Parse(response ?? $"<{TerraConsts.Goods} />", LoadOptions.None);
                 XDocument doc = new(new XElement(TerraConsts.Response, xml.Root));